+++ siga-web/src/main/java/org/itcgae/siga/scs/services/impl/guardia/AsistenciaServiceImpl.java
@@ -5058,7 +5058,7 @@ public class AsistenciaServiceImpl implements AsistenciaService {
 		}
 		scsEjg.setNumerodiligencia(scsAsistencia.getNumerodiligencia());
 		scsEjg.setNig(scsAsistencia.getNig());
-		scsEjg.setIdpretension(scsAsistencia.getIdpretension());
+		scsEjg.setIdpretension(scsAsistencia.getIdpretension().longValue());
 		scsEjg.setUsumodificacion(usuario);
 		scsEjg.setFechamodificacion(new Date());
